+++ acme/resources/CNT_OL.py
@@ -30,6 +30,8 @@ class CNT_OL(Resource):
 		if L.isDebug: L.logDebug('Retrieving oldest CIN from CNT')
 		if (r := self._getOldest()) is None:
 			return Result(rsc=RC.notFound, dbg='no instance for <oldest>')
+		if not (res := r.willBeRetrieved()).status:
+			return res
 		return Result(resource=r)