+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [CSE] Added support for \<ACP>'s *acr/acod* attribute (not for specializations yet, though).
 - [CSE] Added support for 'Request Expiration Timestamp' request parameter, also for \<request> resources.
 - [CSE] Added support for *Vendor Information* request/response header.
+- [CSE] Added support for \<container> *disableRetrieval* attribute.
 - [WEB] Added OAuth2 authorization support for the proxied CSE (for the stand-alone web UI).
 - [WEB] Added opening the web UI in a browser on startup (for the stand-alone web UI).
 - [TESTS] Added OAuth2 authorization support for the tests.